requirements.txt
fastapi==0.121.0
uvicorn==0.38.0
psycopg2==2.9.11
python-dotenv==1.2.1
pydantic==2.12.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20

